# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Goat Nutrition App API
  version: 0.1.0
  description: |
    Minimal OpenAPI for key endpoints. Auth uses cookie session (Flask-Login).
servers:
  - url: http://localhost:5001
paths:
...
  /api/iot/devices:
    get:
      summary: List IoT devices for current user
      security:
        - cookieAuth: []
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
    post:
      summary: Register a new IoT device
      description: |
        Returns device metadata and a one-time `api_key` field. The API key is only returned in this response and
        cannot be retrieved afterwards.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201': { description: Created with one-time API key }
        '400': { description: Validation error }
        '401': { description: Unauthorized }
  /api/iot/devices/{device_id}:
    parameters:
      - in: path
        name: device_id
        required: true
        schema: { type: integer }
    get:
      summary: Retrieve IoT device details
      security:
        - cookieAuth: []
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
        '404': { description: Not Found }
    put:
      summary: Update IoT device metadata
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200': { description: Updated }
        '400': { description: Validation error }
        '401': { description: Unauthorized }
        '404': { description: Not Found }
    delete:
      summary: Remove IoT device
      security:
        - cookieAuth: []
      responses:
        '200': { description: Deleted }
        '401': { description: Unauthorized }
        '404': { description: Not Found }
  /api/iot/ingest:
    post:
      summary: Ingest sensor data from IoT device
      description: Requires `X-API-Key` header signed with device secret. Returns 201 on success.
      parameters:
        - in: header
          name: X-API-Key
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                data:
                  type: object
      responses:
        '201': { description: Accepted and enqueued }
        '400': { description: Invalid payload }
        '401': { description: Missing or invalid API key }
  /api/iot/rules:
    get:
      summary: List automation rules
      security:
        - cookieAuth: []
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
    post:
      summary: Create automation rule
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201': { description: Created }
        '400': { description: Validation error }
        '401': { description: Unauthorized }
  /api/iot/rules/{rule_id}:
    parameters:
      - in: path
        name: rule_id
        required: true
        schema: { type: integer }
    put:
      summary: Update automation rule
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200': { description: Updated }
        '400': { description: Validation error }
        '401': { description: Unauthorized }
        '404': { description: Not Found }
    delete:
      summary: Delete automation rule
      security:
        - cookieAuth: []
      responses:
        '200': { description: Deleted }
        '401': { description: Unauthorized }
        '404': { description: Not Found }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
...
security:
  - cookieAuth: []